Introduce AnnotationManagerDelegate protocol
Now a client of the annoation manager can simply pass a scene view and when it calls add annoations, the annotation manager will attempt to place anchors at the correct locations and provide a default node to visualize them. If the client implements the new node(for anchor) protocol, it can customize the node used for visualization.